% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlighter-pkg.R
\docType{package}
\name{sqlighter}
\alias{sqlighter}
\alias{sqlighter-package}
\title{sqlighter: A package for computating the notorious bar statistic.}
\description{
The sqlighter package provides convenient wrappers that safely open and close
the connection to the SQLite database before and after sending queries and
receiving results.
}
