% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{sqliteListFields}
\alias{sqliteListFields}
\title{List fields available in database table}
\usage{
sqliteListFields(file, table, ...)
}
\arguments{
\item{file}{Path to the SQLite database file}

\item{table}{Name of the table to inspect}

\item{...}{Other parameters passed on to methods.}
}
\value{
A character vector of field names.
}
\description{
List fields available in database table
}
\examples{
# Setup ----

library(RSQLite)

database_file <- tempfile()
conn <- dbConnect(SQLite(), dbname = database_file)
dbWriteTable(conn, "mtcars", mtcars[1:5, ])
dbWriteTable(conn, "USArrests", USArrests[1:5, ])
dbDisconnect(conn)

# Example ----

sqliteListFields(database_file, "mtcars")
}
