% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{sqliteListTables}
\alias{sqliteListTables}
\title{List table names in an SQLite database}
\usage{
sqliteListTables(file)
}
\arguments{
\item{file}{Path to the SQLite database file}
}
\value{
A character vector of table names.
}
\description{
List table names in an SQLite database
}
\examples{
# Setup ----
library(RSQLite)

database_file <- tempfile()
conn <- dbConnect(SQLite(), dbname = database_file)
dbWriteTable(conn, "mtcars", mtcars[1:5, ])
dbWriteTable(conn, "USArrests", USArrests[1:5, ])
dbDisconnect(conn)

# Example ----

sqliteListTables(database_file)
}
