% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsqlite.R
\name{sqliteTableExists}
\alias{sqliteTableExists}
\title{Check if a table exists in the database}
\usage{
sqliteTableExists(table, file)
}
\arguments{
\item{table}{Name of the table to check for existence}

\item{file}{Database to query}
}
\value{
A logical value that indicates whether the table exists in the
database.
}
\description{
Check if a table exists in the database
}
\examples{
# Example data ----
library(RSQLite)

database_file <- tempfile()
conn <- dbConnect(SQLite(), dbname = database_file)
dbWriteTable(conn, "mtcars", mtcars[1:5, ])
dbWriteTable(conn, "USArrests", USArrests[1:5, ])
dbDisconnect(conn)


# Example ----

sqliteTableExists("status_summary", database_file)
}
